--- services/index/internal/adapter/mongodb/mongodb.go
@@ -1,20 +1,0 @@
-package mongodb
-
-import (
-	"github.com/MurmurationsNetwork/MurmurationsServices/common/logger"
-	"github.com/MurmurationsNetwork/MurmurationsServices/common/mongo"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/config"
-)
-
-func Init() {
-	uri := mongo.GetURI(config.Conf.Mongo.USERNAME, config.Conf.Mongo.PASSWORD, config.Conf.Mongo.HOST)
-
-	err := mongo.NewClient(uri, config.Conf.Mongo.DBName)
-	if err != nil {
-		logger.Panic("error when trying to connect to MongoDB", err)
-	}
-	err = mongo.Client.Ping()
-	if err != nil {
-		logger.Panic("error when trying to ping the MongoDB", err)
-	}
-}
