+++ services/index/internal/app/application.go
@@ -6,17 +6,12 @@ import (
 
 	"github.com/MurmurationsNetwork/MurmurationsServices/common/logger"
 	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/config"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/adapter/elasticsearch"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/adapter/mongodb"
-	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/internal/adapter/nats"
+	"github.com/MurmurationsNetwork/MurmurationsServices/services/index/global"
 	"github.com/gin-gonic/gin"
 )
 
 func init() {
-	config.Init()
-	elasticsearch.Init()
-	mongodb.Init()
-	nats.Init()
+	global.Init()
 }
 
 func StartApplication() {
